Add Serilog structured logging to Penjualan and BackOffice
- Add Serilog with File, Console, and Debug sinks to both projects
- Configure rolling daily log files with 30-day retention
- Add global exception handlers (ThreadException + UnhandledException)
- Migrate BackOffice DatabaseHelper from hardcoded to appsettings.json
- Replace Console.WriteLine/Debug.WriteLine with structured Log calls
- Upgrade Microsoft.Extensions.Configuration to 10.0.5

+++ BackOffice/Model/DTOPinjaman.cs
@@ -12,7 +12,7 @@ namespace BackOffice.Model
         public DateTime TANGGAL { get; set; }
         public string JAM { get; set; }
         public string KASIR { get; set; }
-        public double ID_PELANGGAN { get; set; }
+        public int ID_PELANGGAN { get; set; }
         public string NIK { get; set; }
         public string NAMA_PELANGGAN { get; set; }
         public string STATUS { get; set; }
